refactor: file organization + multi-device debugging config
